refactor dino wrapper
